Refactor code locations and first attempt at user provisioning
