+++ tracker/view/login.py
@@ -9,7 +9,7 @@ from werkzeug.exceptions import Unauthorized
 from config import TRACKER_PASSWORD_LENGTH_MAX
 from config import TRACKER_PASSWORD_LENGTH_MIN
 from config import SSO_ENABLED
-from tracker import tracker
+from tracker import tracker, oauth
 from tracker.form import LoginForm
 from tracker.model.user import User
 from tracker.user import user_assign_new_token
@@ -52,4 +52,39 @@ def logout():
 
     user_invalidate(current_user)
     logout_user()
+    return redirect(url_for('tracker.index'))
+
+
+@tracker.route('/sso-auth')
+def sso_auth():
+    from tracker import db
+
+    # login the user here, create a session and set role
+    token = oauth.idp.authorize_access_token()
+    parsed_token = oauth.idp.parse_id_token(token)
+    # check if user can be matched against local db of users
+    user = db.get(User, email=parsed_token.get('email'))
+
+    if user:
+        user = user_assign_new_token(user)
+        user.is_authenticated = True
+        login_user(user)
+    elif len(parsed_token.get('roles')) == 0:
+        return redirect(url_for('tracker.index'))
+    else:
+        # user does not exist in local db
+        # need to create him to leverage existing user access controls
+        from tracker.user import random_string, hash_password
+
+        user = User()
+        user.name = parsed_token.get('preferred_username')
+        user.email = parsed_token.get('email')
+        user.salt = random_string()
+        user.password = hash_password('wasd', user.salt)
+        user.role = parsed_token.get('roles')[0] 
+        user.active = True
+
+        db.session.add(user)
+        db.session.commit()
+
     return redirect(url_for('tracker.index'))
